# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Test Server API",
    "version": "1.0.0",
    "description": "API for testing API Gateway integration"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    },
    {
      "url": "https://salem-processors-rhythm-brought.trycloudflare.com",
      "description": "Public server (via Cloudflare)"
    }
  ],
  "paths": {
    "/test": {
      "get": {
        "summary": "Test GET endpoint",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Test POST endpoint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login to get JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT Token" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/profile": {
      "get": {
        "summary": "Protected profile endpoint",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/booking": {
      "post": {
        "summary": "Protected booking endpoint",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Booking response" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}